Goal 1- Refactor repeated code to use super class
