+++ src/test/java/chess/pieces/RookTest.java
@@ -19,7 +19,7 @@ public class RookTest {
 
     private GameState gameState;
     private Map<Piece, Position> board;
-    private Piece rook;
+    private Rook rook;
     private String testPosition;
     private Set<String> resultMoves;
 
@@ -33,7 +33,7 @@ public class RookTest {
     public void testGetPossibleMovesGameInit() {
         testPosition = "a1";
         gameState.reset();
-        rook = gameState.getPieceAt(testPosition);
+        rook = (Rook) gameState.getPieceAt(testPosition);
         resultMoves = rook.getPossibleMoves(gameState.getPositionToPieceMap());
         assertEquals(0, resultMoves.size());
     }
@@ -43,7 +43,7 @@ public class RookTest {
         testPosition = "c1";
         board.put(new Rook(Player.White), new Position(testPosition));
         gameState.setFakeBoard(board);
-        rook = gameState.getPieceAt(testPosition);
+        rook = (Rook) gameState.getPieceAt(testPosition);
         resultMoves = rook.getPossibleMoves(gameState.getPositionToPieceMap());
         assertEquals(14, resultMoves.size());
     }
@@ -53,7 +53,7 @@ public class RookTest {
         testPosition = "d5";
         board.put(new Rook(Player.Black), new Position(testPosition));
         gameState.setFakeBoard(board);
-        rook = gameState.getPieceAt(testPosition);
+        rook = (Rook) gameState.getPieceAt(testPosition);
         resultMoves = rook.getPossibleMoves(gameState.getPositionToPieceMap());
         assertEquals(14, resultMoves.size());
     }
@@ -69,7 +69,7 @@ public class RookTest {
         board.put(new Knight(Player.White), new Position("h5"));
         gameState.setFakeBoard(board);
 
-        rook = gameState.getPieceAt(testPosition);
+        rook = (Rook) gameState.getPieceAt(testPosition);
         resultMoves = rook.getPossibleMoves(gameState.getPositionToPieceMap());
         assertEquals(7, resultMoves.size());
     }
